pyproject.toml
[project]
name = "minder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "mcp>=1.26.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings[toml]>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy[asyncio]>=2.0.48",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]
